requirements.txt
websockets
requests
pycryptodome
